move parseAddExternalToolInput out of  ExternalToolHandler #4230
Also add FIXMEs based on code review.
