+++ src/main/java/edu/harvard/iq/dataverse/externaltools/ExternalToolServiceBean.java
@@ -35,6 +35,7 @@ public class ExternalToolServiceBean {
 
     public ExternalTool save(ExternalTool externalTool) {
         em.persist(externalTool);
+        // FIXME: Remove this flush.
         em.flush();
         return em.merge(externalTool);
     }
