+++ src/test/java/edu/harvard/iq/dataverse/externaltools/ExternalToolHandlerTest.java
@@ -2,33 +2,11 @@ package edu.harvard.iq.dataverse.externaltools;
 
 import edu.harvard.iq.dataverse.DataFile;
 import edu.harvard.iq.dataverse.authorization.users.ApiToken;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import javax.json.Json;
 import org.junit.Test;
 import static org.junit.Assert.assertEquals;
 
 public class ExternalToolHandlerTest {
-
-    @Test
-    public void testParseAddExternalToolInput() throws Exception {
-        assertEquals(null, ExternalToolHandler.parseAddExternalToolInput(null));
-        assertEquals(null, ExternalToolHandler.parseAddExternalToolInput(""));
-        assertEquals(null, ExternalToolHandler.parseAddExternalToolInput(Json.createObjectBuilder().build().toString()));
-        String psiTool = new String(Files.readAllBytes(Paths.get("doc/sphinx-guides/source/_static/installation/files/root/external-tools/psi.json")));
-        System.out.println("psiTool: " + psiTool);
-        ExternalTool externalTool = ExternalToolHandler.parseAddExternalToolInput(psiTool);
-        DataFile dataFile = new DataFile();
-        dataFile.setId(42l);
-        externalTool.setDataFile(dataFile);
-        ApiToken apiToken = new ApiToken();
-        apiToken.setTokenString("7196b5ce-f200-4286-8809-03ffdbc255d7");
-        externalTool.setApiToken(apiToken);
-        String toolUrl = externalTool.getToolUrl();
-        System.out.println("result: " + toolUrl);
-        assertEquals("https://beta.dataverse.org/custom/DifferentialPrivacyPrototype/UI/code/interface.html?fileid=42&key=7196b5ce-f200-4286-8809-03ffdbc255d7", toolUrl);
-
-    }
 
     @Test
     public void testGetToolUrlWithOptionalQueryParameters() {
